# backend/ms-cv/src/main/java/uasz/alumni/ms_cv/controllers/ComptencesApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.17.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package uasz.alumni.ms_cv.controllers;

import uasz.alumni.ms_cv.models.Competence;
import uasz.alumni.ms_cv.models.CompetenceCreate;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-26T19:36:29.996620853Z[Africa/Dakar]", comments = "Generator version: 7.17.0")
@Validated
@Tag(name = "Compétences", description = "the Compétences API")
public interface ComptencesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_COMPETENCES_ID_DELETE = "/competences/{id}";
    /**
     * DELETE /competences/{id} : Supprimer une compétence
     *
     * @param id  (required)
     * @return Compétence supprimée (status code 204)
     */
    @Operation(
        operationId = "competencesIdDelete",
        summary = "Supprimer une compétence",
        tags = { "Compétences" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Compétence supprimée")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = ComptencesApi.PATH_COMPETENCES_ID_DELETE
    )
    default ResponseEntity<Void> competencesIdDelete(
        @NotNull @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_COMPETENCES_ID_PUT = "/competences/{id}";
    /**
     * PUT /competences/{id} : Modifier une compétence
     *
     * @param id  (required)
     * @param competenceCreate  (required)
     * @return Compétence mise à jour (status code 200)
     */
    @Operation(
        operationId = "competencesIdPut",
        summary = "Modifier une compétence",
        tags = { "Compétences" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Compétence mise à jour")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = ComptencesApi.PATH_COMPETENCES_ID_PUT,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> competencesIdPut(
        @NotNull @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "CompetenceCreate", description = "", required = true) @Valid @RequestBody CompetenceCreate competenceCreate
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_COMPETENCES_POST = "/competences";
    /**
     * POST /competences : Ajouter une compétence
     *
     * @param competenceCreate  (required)
     * @return Compétence ajoutée (status code 201)
     */
    @Operation(
        operationId = "competencesPost",
        summary = "Ajouter une compétence",
        tags = { "Compétences" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Compétence ajoutée", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Competence.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ComptencesApi.PATH_COMPETENCES_POST,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Competence> competencesPost(
        @Parameter(name = "CompetenceCreate", description = "", required = true) @Valid @RequestBody CompetenceCreate competenceCreate
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"cvId\" : 1, \"categorie\" : \"Langage de programmation\", \"id\" : 1, \"nom\" : \"Java\", \"niveau\" : \"AVANCE\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}