# backend/ms-cv/src/main/java/uasz/alumni/ms_cv/controllers/ExpriencesApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.17.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package uasz.alumni.ms_cv.controllers;

import uasz.alumni.ms_cv.models.Experience;
import uasz.alumni.ms_cv.models.ExperienceCreate;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-26T19:36:29.996620853Z[Africa/Dakar]", comments = "Generator version: 7.17.0")
@Validated
@Tag(name = "Expériences", description = "the Expériences API")
public interface ExpriencesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_EXPERIENCES_ID_DELETE = "/experiences/{id}";
    /**
     * DELETE /experiences/{id} : Supprimer une expérience
     *
     * @param id  (required)
     * @return Expérience supprimée (status code 204)
     */
    @Operation(
        operationId = "experiencesIdDelete",
        summary = "Supprimer une expérience",
        tags = { "Expériences" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Expérience supprimée")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = ExpriencesApi.PATH_EXPERIENCES_ID_DELETE
    )
    default ResponseEntity<Void> experiencesIdDelete(
        @NotNull @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_EXPERIENCES_ID_PUT = "/experiences/{id}";
    /**
     * PUT /experiences/{id} : Modifier une expérience
     *
     * @param id  (required)
     * @param experienceCreate  (required)
     * @return Expérience mise à jour (status code 200)
     */
    @Operation(
        operationId = "experiencesIdPut",
        summary = "Modifier une expérience",
        tags = { "Expériences" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Expérience mise à jour", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Experience.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = ExpriencesApi.PATH_EXPERIENCES_ID_PUT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Experience> experiencesIdPut(
        @NotNull @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "ExperienceCreate", description = "", required = true) @Valid @RequestBody ExperienceCreate experienceCreate
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"cvId\" : 1, \"entreprise\" : \"Tech Company\", \"dateDebut\" : \"2020-01-15\", \"localisation\" : \"Dakar, Sénégal\", \"description\" : \"- Développement d'applications web\n- Gestion de bases de données\", \"id\" : 1, \"poste\" : \"Développeur Full Stack\", \"dateFin\" : \"2023-06-30\", \"enCours\" : false }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_EXPERIENCES_POST = "/experiences";
    /**
     * POST /experiences : Ajouter une expérience professionnelle
     *
     * @param experienceCreate  (required)
     * @return Expérience ajoutée (status code 201)
     *         or Données invalides (status code 400)
     */
    @Operation(
        operationId = "experiencesPost",
        summary = "Ajouter une expérience professionnelle",
        tags = { "Expériences" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Expérience ajoutée", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Experience.class))
            }),
            @ApiResponse(responseCode = "400", description = "Données invalides")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ExpriencesApi.PATH_EXPERIENCES_POST,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Experience> experiencesPost(
        @Parameter(name = "ExperienceCreate", description = "", required = true) @Valid @RequestBody ExperienceCreate experienceCreate
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"cvId\" : 1, \"entreprise\" : \"Tech Company\", \"dateDebut\" : \"2020-01-15\", \"localisation\" : \"Dakar, Sénégal\", \"description\" : \"- Développement d'applications web\n- Gestion de bases de données\", \"id\" : 1, \"poste\" : \"Développeur Full Stack\", \"dateFin\" : \"2023-06-30\", \"enCours\" : false }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}