# backend/ms-cv/src/main/java/uasz/alumni/ms_cv/controllers/GnrationCvApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.17.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package uasz.alumni.ms_cv.controllers;

import org.springframework.lang.Nullable;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-26T19:36:29.996620853Z[Africa/Dakar]", comments = "Generator version: 7.17.0")
@Validated
@Tag(name = "Génération CV", description = "the Génération CV API")
public interface GnrationCvApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_CV_ID_GENERER_PDF_GET = "/cv/{id}/generer-pdf";
    /**
     * GET /cv/{id}/generer-pdf : Générer le CV en PDF standardisé
     * Génère un PDF du CV avec le template UASZ
     *
     * @param id  (required)
     * @param template Template de CV à utiliser (optional, default to MODERNE)
     * @return PDF généré avec succès (status code 200)
     *         or CV non trouvé (status code 404)
     */
    @Operation(
        operationId = "cvIdGenererPdfGet",
        summary = "Générer le CV en PDF standardisé",
        description = "Génère un PDF du CV avec le template UASZ",
        tags = { "Génération CV" },
        responses = {
            @ApiResponse(responseCode = "200", description = "PDF généré avec succès", content = {
                @Content(mediaType = "application/pdf", schema = @Schema(implementation = org.springframework.core.io.Resource.class))
            }),
            @ApiResponse(responseCode = "404", description = "CV non trouvé")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = GnrationCvApi.PATH_CV_ID_GENERER_PDF_GET,
        produces = { "application/pdf" }
    )
    default ResponseEntity<org.springframework.core.io.Resource> cvIdGenererPdfGet(
        @NotNull @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "template", description = "Template de CV à utiliser", in = ParameterIn.QUERY) @Valid @RequestParam(value = "template", required = false, defaultValue = "MODERNE") String template
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_CV_ID_PREVISUALISER_GET = "/cv/{id}/previsualiser";
    /**
     * GET /cv/{id}/previsualiser : Prévisualiser le CV avant génération
     * Génère une prévisualisation HTML du CV
     *
     * @param id  (required)
     * @param template  (optional, default to MODERNE)
     * @return Prévisualisation générée (status code 200)
     *         or CV non trouvé (status code 404)
     */
    @Operation(
        operationId = "cvIdPrevisualiserGet",
        summary = "Prévisualiser le CV avant génération",
        description = "Génère une prévisualisation HTML du CV",
        tags = { "Génération CV" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Prévisualisation générée", content = {
                @Content(mediaType = "text/html", schema = @Schema(implementation = String.class))
            }),
            @ApiResponse(responseCode = "404", description = "CV non trouvé")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = GnrationCvApi.PATH_CV_ID_PREVISUALISER_GET,
        produces = { "text/html" }
    )
    default ResponseEntity<String> cvIdPrevisualiserGet(
        @NotNull @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "template", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "template", required = false, defaultValue = "MODERNE") String template
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}