# openapi/ms-cv.yaml
openapi: 3.0.3
info:
  title: CV Service API
  version: 1.0.0
  description: API de gestion des CV des alumni et étudiants - Génération de CV standardisés
servers:
  - url: http://localhost:8082/api/v1/cv
paths:
...
  /cv/{id}/generer-pdf:
    get:
      tags:
        - Génération CV
      summary: Générer le CV en PDF standardisé
      description: Génère un PDF du CV avec le template UASZ
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: template
          in: query
          schema:
            type: string
            enum: [MODERNE, CLASSIQUE, PROFESSIONNEL]
            default: MODERNE
          description: Template de CV à utiliser
      responses:
        "200":
          description: PDF généré avec succès
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          description: CV non trouvé
  /cv/{id}/previsualiser:
    get:
      tags:
        - Génération CV
      summary: Prévisualiser le CV avant génération
      description: Génère une prévisualisation HTML du CV
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: template
          in: query
          schema:
            type: string
            enum: [MODERNE, CLASSIQUE, PROFESSIONNEL]
            default: MODERNE
      responses:
        "200":
          description: Prévisualisation générée
          content:
            text/html:
              schema:
                type: string
        "404":
          description: CV non trouvé
  /experiences:
    post:
      tags:
        - Expériences
      summary: Ajouter une expérience professionnelle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperienceCreate"
      responses:
        "201":
          description: Expérience ajoutée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"
        "400":
          description: Données invalides
  /experiences/{id}:
    put:
      tags:
        - Expériences
      summary: Modifier une expérience
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperienceCreate"
      responses:
        "200":
          description: Expérience mise à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"

    delete:
      tags:
        - Expériences
      summary: Supprimer une expérience
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Expérience supprimée
...
  /competences:
    post:
      tags:
        - Compétences
      summary: Ajouter une compétence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetenceCreate"
      responses:
        "201":
          description: Compétence ajoutée
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competence"
  /competences/{id}:
    put:
      tags:
        - Compétences
      summary: Modifier une compétence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetenceCreate"
      responses:
        "200":
          description: Compétence mise à jour

    delete:
      tags:
        - Compétences
      summary: Supprimer une compétence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Compétence supprimée
...
components:
  schemas:
...
    Experience:
      type: object
      properties:
        id:
          type: integer
          example: 1
        cvId:
          type: integer
          example: 1
        poste:
          type: string
          example: "Développeur Full Stack"
        entreprise:
          type: string
          example: "Tech Company"
        localisation:
          type: string
          example: "Dakar, Sénégal"
        dateDebut:
          type: string
          format: date
          example: "2020-01-15"
        dateFin:
          type: string
          format: date
          example: "2023-06-30"
        enCours:
          type: boolean
          example: false
        description:
          type: string
          example: "- Développement d'applications web\n- Gestion de bases de données"
    ExperienceCreate:
      type: object
      required:
        - cvId
        - poste
        - entreprise
        - dateDebut
      properties:
        cvId:
          type: integer
        poste:
          type: string
        entreprise:
          type: string
        localisation:
          type: string
        dateDebut:
          type: string
          format: date
        dateFin:
          type: string
          format: date
        enCours:
          type: boolean
          default: false
        description:
          type: string
...
    Competence:
      type: object
      properties:
        id:
          type: integer
          example: 1
        cvId:
          type: integer
          example: 1
        nom:
          type: string
          example: "Java"
        niveau:
          type: string
          enum: [DEBUTANT, INTERMEDIAIRE, AVANCE, EXPERT]
          example: "AVANCE"
        categorie:
          type: string
          example: "Langage de programmation"
    CompetenceCreate:
      type: object
      required:
        - cvId
        - nom
      properties:
        cvId:
          type: integer
        nom:
          type: string
        niveau:
          type: string
          enum: [DEBUTANT, INTERMEDIAIRE, AVANCE, EXPERT]
          default: INTERMEDIAIRE
        categorie:
          type: string
...